Update and rename Data.java to Most-frequency-number in sorted array.java
The sorted array is to be either in ascending or descending .
Firstly enter the size of array then enter your data it must be sorted and numeric. Then you will get output as most frequent number.
For example :  input -  9 9 7 7 2 2 2 
                        output- 2
In above example 9 will occur in 2 times ,7 will occur in 2 times and 2 occurs 3 times. So the output will be the 2.